requirements.txt
asgiref==3.11.0
Django>=5.2,<5.3
sqlparse==0.5.5
tzdata==2025.3
gunicorn==21.2.0
psycopg2-binary==2.9.10
djangorestframework~=3.16.0
dj-rest-auth==7.0.0
django-allauth==65.2.0
djangorestframework-simplejwt~=5.5.1
django-silk
python-json-logger~=3.3.0
drf-spectacular~=0.28.0
requests==2.32.3
django-structlog==10.0.0